Picture upload [Should] fully work, at least with Postman requests.
First the client posts the meta data to the server, then the server returns an upload link, then the user multiparts the image[s] to the server at this link. Right now I can only verify it working for one picture at  a time. Which fits the use case as people will take photos with tthis the get it into the system at first. Later on I will add storage upload, and getting pictures that have already been taken imported into Pix. But that introduces a whole lot of variables I don;t need to worry about at the moment. And there are implementation locations to place those changes. But they would mostly be client side...